requirements.txt
hydra-core
pytorch-lightning
torchvision
wandb
jupyterlab
numpy 
matplotlib
torch
pillow
scikit-learn